feat: move whitelist to decorators
