+++ frappe/__init__.py
@@ -47,6 +47,9 @@ from frappe.query_builder.utils import (
 from frappe.utils.caching import request_cache
 from frappe.utils.data import cint, cstr, sbool
 
+# Backward compatibility
+from .decorators import whitelist
+
 # Local application imports
 from .exceptions import *
 from .types import Filters, FilterSignature, FilterTuple, _dict
@@ -780,55 +783,6 @@ whitelisted: set[Callable] = set()
 guest_methods: set[Callable] = set()
 xss_safe_methods: set[Callable] = set()
 allowed_http_methods_for_whitelisted_func: dict[Callable, list[str]] = {}
-
-
-def whitelist(allow_guest=False, xss_safe=False, methods=None):
-	"""
-	Decorator for whitelisting a function and making it accessible via HTTP.
-	Standard request will be `/api/method/[path.to.method]`
-
-	:param allow_guest: Allow non logged-in user to access this method.
-	:param methods: Allowed http method to access the method.
-
-	Use as:
-
-	        @frappe.whitelist()
-	        def myfunc(param1, param2):
-	                pass
-	"""
-
-	if not methods:
-		methods = ["GET", "POST", "PUT", "DELETE"]
-
-	def innerfn(fn):
-		from frappe.utils.typing_validations import validate_argument_types
-
-		global whitelisted, guest_methods, xss_safe_methods, allowed_http_methods_for_whitelisted_func
-
-		# validate argument types only if request is present
-		in_request_or_test = lambda: getattr(local, "request", None) or local.flags.in_test  # noqa: E731
-
-		# get function from the unbound / bound method
-		# this is needed because functions can be compared, but not methods
-		method = None
-		if hasattr(fn, "__func__"):
-			method = validate_argument_types(fn, apply_condition=in_request_or_test)
-			fn = method.__func__
-		else:
-			fn = validate_argument_types(fn, apply_condition=in_request_or_test)
-
-		whitelisted.add(fn)
-		allowed_http_methods_for_whitelisted_func[fn] = methods
-
-		if allow_guest:
-			guest_methods.add(fn)
-
-			if xss_safe:
-				xss_safe_methods.add(fn)
-
-		return method or fn
-
-	return innerfn
 
 
 def is_whitelisted(method):
